+++ README.md
@@ -1,12 +1,12 @@
 # What?
 
-This package creates a binary that is built from *cmd/main.go
+This package creates binaries that are built from *cmd*/main.go
 
-This binary can be zipped into a .zip file and then deployed as a lambda function
+These binaries can be zipped into .zip files and then deployed as lambda functions.
 
 # Description
 
-When using AWS Elastic Disaster Recovery Service (DRS) every source machine has a corresponding launch templates. Launch templates cannot be edited in a batch using the native DRS tooling. This package is a lambda function that allows for editing a group of launch templates that all are tagged with a certain key in the DRS console.
+When using AWS Elastic Disaster Recovery Service (DRS) every source machine has a corresponding launch template. Launch templates cannot be edited in a batch using the native DRS tooling. This package is a set of lambda functions that allow for editing a group of launch templates that all are tagged with a certain key in the DRS console.
 
 For Example:
 - Create one launch template for all servers tagged 'DB'.
@@ -15,19 +15,29 @@ For Example:
 
 - Create one launch template that applies to all replicating DRS servers.
 
+# The Architecture
+
 ![templatearch](https://user-images.githubusercontent.com/97046295/161995172-a5a3475e-319e-480a-9bb2-c1f32338e89b.png)
+
+This solution is composed of the following components:
+
+- S3 bucket for storing launch templates in the form of json files.
+
+- Lambda function that pulls down a json launch template from the bucket and then updates DRS servers that are tagged with the prefix of that json. This function will be called 'set-drs-templates'.
+
+- Lambda function that runs on a schedule and scans for any new replicating servers that have a matching tag to one of the existing templates in the bucket. This allows new servers that are added to DRS to inherit the launch template that they are tagged for. In this example we are going to just set the scheduler to run once daily to capture new servers added throughout the day. This function will be called 'schedule-drs-templates'.
 
 # Prerequisites
 
 In order to use this solution, it is required to have actively replicating servers in DRS. For more information on getting started with DRS reference the [quick start guide](https://docs.aws.amazon.com/drs/latest/userguide/getting-started.html).
 
-Part of this solution is creating a lambda function which needs to make API calls to both DRS and EC2. It is required to have a role with the proper permissions to access both. You can create a role with 3 managed policies:
+Part of this solution is creating a lambda function which needs to make API calls to  DRS, EC2, and S3. It is required to have a role with the proper permissions to access all three services. You can create a role with 3 managed policies for simplicity:
 
 - "AWSElasticDisasterRecoveryReadOnlyAccess"
 
 - "AmazonEC2FullAccess"
 
-- "AmazonS3ReadOnlyAccess"
+- "AmazonS3FullAccess"
 
 Here is a full list of API calls made by the solution if you would like to create a more granular policy:
 
@@ -37,38 +47,80 @@ Here is a full list of API calls made by the solution if you would like to creat
 
 - "s3.GetObject"
 
+- "s3.PutObject"
+
 - "ec2.CreateLaunchTemplateVersion"
 
 - "ec2.ModifyLaunchTemplate"
 
 # Usage
 
-Create the Lambda Function:
+Deploying the solution is composed of three main steps. Create the lambda functions, Create the S3 bucket trigger, and Creating a template.
 
-- Clone the repo
+Create the Lambda Functions:
+
+* Clone the repo
 ```
 git clone https://github.com/aws-samples/drs-template-manager.git
 ```
 
-- Create the zip
+* Create the zip deployment package of the 'set-drs-templates' function.
 ```
 cd drs-template-manager
 cd cmd
-zip function.zip drs-template-manager
+zip template.zip drs-template-manager
 ```
 
-- Make a new GO lambda function in the same region as your DRS replicating servers and use the 'function.zip' as the deployment package. Under "Runtime Settings" Set the Handler to 'drs-template-manager'. The architecture should be x86 and the Runtime should be Go 1.x .
+* Create the zip deployment package of the 'schedule-drs-templates' function.
+```
+cd ../cmd-cron
+zip cron.zip template-cron-automation
+```
+
+- Make two new GO lambda function in the same region as your DRS replicating servers and use the '.zip' files created above as the deployment packages. Under "Runtime Settings" Set the Handler to 'drs-template-manager' for the function that sets the templates and 'template-cron-automation' for the scheduler. The architecture should be x86 and the Runtime should be Go 1.x .
+
+* Create the 'schedule-drs-templates' function, replace '$INSERTROLEARN' with the arn of the role you created for the solution.
 ```
 aws lambda create-function \            
---function-name sandboxGoTemplate \
+--function-name schedule-drs-templates \
+--role $INSERTROLEARN
+--runtime go1.x \
+--handler template-cron-automation \
+--package-type Zip \
+--zip-file fileb://cron.zip
+```
+
+* Create the 'set-drs-templates' function, replace '$INSERTROLEARN' with the arn of the role you created for the solution.
+```
+cd ../cmd-template
+aws lambda create-function \            
+--function-name set-drs-templates \
 --role $INSERTROLEARN
 --runtime go1.x \
 --handler drs-template-manager \
 --package-type Zip \
---zip-file fileb://function.zip
+--zip-file fileb://template.zip
 ```
 
-Create the S3 bucket trigger:
+- Once the scheduler is created you need to determine how often you would like it to run and then create a CloudWatch cron event to trigger it. For this example we will create an event rule that triggers once per day at 12:00 PM UTC. Once we make the rule it needs to be added to the lambda function as a trigger.
+
+* Create the rule
+```
+aws events put-rule \
+--schedule-expression "cron(0 12 * * ? *)" \
+--name template-cron-rule
+```
+
+* Add the 'schedule-drs-templates' function as a target for the rule. Replace $FunctionARN with the ARN of the 'schedule-drs-templates' lambda function.
+```
+aws events put-targets \
+--rule template-cron-rule \
+--targets "Id"="1","Arn"=$FunctionARN
+```
+
+
+
+Create the S3 bucket trigger for the 'set-drs-templates' function:
 
 - Create an S3 bucket in the same region as the lambda function.
 ```
@@ -99,9 +151,14 @@ For Example:
 
 # Build Locally
 
-You can build this locally if you have the go tooling installed.
+You can build the functions locally if you have the go tooling installed.
 ```
-cd cmd
+cd cmd-cron
+GOOS=linux go build .
+```
+
+```
+cd cmd-template
 GOOS=linux go build .
 ```
